requirements.txt
SpeechRecognition
streamlit_mic_recorder
ffmpeg-python
google-generativeai
pydub 
gTTS 
edge-tts
PyMuPDF
python-docx
reportlab
moviepy
